requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
sqlalchemy2-stubs
psycopg2-binary
python-dotenv
requests
argon2-cffi
jwt



